Connecting to remote server and receiving responses functionality added.
Done to allow the server to make requests to a specified remote server
and receive the response from that server, as specified by the spec.

Changed code in proxy_request_handler to start a socket connection
between the server and an another remote server.

Tested in the web browser, and works as intended.

Change-Id: Ibdfece6240cad7b6b7f36242b0c4f892b097ad53