Modified integration.sh to have test for proxy handler and modified proxy_request_handler to handle server addr instead of localhost.
The integration test is done by opening two servers that one is having proxy handler for the other server. And the reqeust is done calling the reverse proxy for the server who has handler for the other server. We were able to get the desired result when calling the server that connects to the other server.

Change-Id: I135c026bc3317113bb9cf8938eafc4525dbb65eb